methods/deepseek-dspy-agent/pyproject.toml
[project]
name = "deepseek-dspy-agent"
version = "0.1.0"
description = "DeepSeek Reasoner + DSPy GEPA Agent for Spider 2.0 benchmarks"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Spider 2.0 Team"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "dspy>=3.0.4,<4.0.0",
    "openai>=2.14.0,<4.0.0",
    "requests>=2.32.5,<3.0.0",
    "tiktoken>=0.12.0,<1.0.0",
    "numpy>=2.4.0,<3.0.0",
    "pandas>=2.3.3,<3.0.0",
    "tqdm>=4.67.1,<5.0.0",
    "jsonlines>=4.0.0,<5.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
deepseek-dspy-agent = "deepseek_dspy_agent.cli:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["deepseek_dspy_agent*"]
exclude = ["tests", "tests.*"]

# Note: spider-agent-lite is an external dependency not available on PyPI.
# It must be installed separately from the Spider 2.0 repository.